requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.1
binaryornot==0.4.4
celery==5.4.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
cron-descriptor==1.4.5
Django==5.1.7
django-browser-reload==1.18.0
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-nested-admin==4.1.1
django-tailwind==3.8.0
django-timezone-field==7.1
django-unfold==0.51.0
idna==3.10
Jinja2==3.1.6
kombu==5.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
prompt_toolkit==3.0.50
Pygments==2.19.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-monkey-business==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
